Move mathvariant related code to MathVariant https://bugs.webkit.org/show_bug.cgi?id=202302
Reviewed by Frédéric Wang, Tim Nguyen, and Darin Adler.

Part of WebKit#51271. Moves the MathVariant enum and the mathVariant
function to the MathVariant file, making it cleaner to implement
text-transform: math-auto in the next patch.

Moves the convertToSingleCodePoint function to StringView. Now it doesn't
perform whitespace trimming of the input string, as MathML Core doesn't
define this. The two existing callers of this function are changed to do
the trimming beforehand.

No behaviour change.

* Source/WTF/wtf/text/StringView.cpp:
(WTF::StringView::convertToSingleCodePoint const):
* Source/WTF/wtf/text/StringView.h:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/mathml/MathMLElement.h:
* Source/WebCore/mathml/MathMLOperatorElement.cpp:
(WebCore::MathMLOperatorElement::parseOperatorChar):
* Source/WebCore/mathml/MathMLPresentationElement.cpp:
(WebCore::MathMLPresentationElement::parseMathVariantAttribute):
(WebCore::MathMLPresentationElement::specifiedMathVariant):
* Source/WebCore/mathml/MathMLTokenElement.cpp:
(WebCore::MathMLTokenElement::convertToSingleCodePoint): Deleted.
* Source/WebCore/mathml/MathMLTokenElement.h:
* Source/WebCore/rendering/mathml/MathMLStyle.cpp:
(WebCore::MathMLStyle::updateStyleIfNeeded):
(WebCore::MathMLStyle::resolveMathMLStyle):
* Source/WebCore/rendering/mathml/MathMLStyle.h:
* Source/WebCore/rendering/mathml/MathVariant.cpp: Added.
(WebCore::mathVariantMapCodePoint):
* Source/WebCore/rendering/mathml/MathVariant.h: Added.
* Source/WebCore/rendering/mathml/RenderMathMLToken.cpp:
(WebCore::RenderMathMLToken::updateMathVariantGlyph):
(WebCore::ExtractKey): Deleted.
(WebCore::MathVariantMappingSearch): Deleted.
(WebCore::mathVariant): Deleted.
* Source/WebCore/rendering/svg/RenderSVGContainer.cpp:
* Source/WebCore/rendering/svg/RenderSVGImage.cpp:
* Source/WebCore/svg/SVGGraphicsElement.cpp:

Canonical link: https://commits.webkit.org/301644@main
